backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
sqlalchemy
pydantic
llama-index
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
sentence-transformers
pypdf
python-docx
python-pptx
reportlab